DCP-084
Did a bit, but it's not entirely working.
It can 'accurately' number how many cells have land, but the specific cells it lists as being 'land' is incorrect.

I also need to find a way to identify  'islands' or 'groups of land'.
I thought I had a method with iterating Explore(), but it ended with StackOverFlow Exceptions.
So yeah, idk.

It's a fun problem, but man it's tricky :P